rename updater package to be main_loop
